+++ src/webparts/restTester/components/IKnownApis.ts
@@ -1,0 +1,8 @@
+export interface IKnownAPIs {
+  api: IApi[];
+}
+
+export interface IApi {
+  method: string;
+  url: string;
+}