+++ src/webparts/restTester/components/apiSuggestions/IApiSuggestionsProps.ts
@@ -1,0 +1,6 @@
+export interface IApiSuggestionsProps {
+  inputVal: string;
+  method: string;
+
+  fChangeApiUrl: (apiUrl: string) => void;
+}