+++ src/webparts/restTester/components/apiSuggestions/IApiSuggestionsState.ts
@@ -1,0 +1,7 @@
+import { IApi } from "../IKnownApis";
+
+export interface IApiSuggestionsState {
+  apiUrls: IApi[];
+  apiBegin: string;
+  apiEnd: string;
+}