+++ src/webparts/restTester/components/apiSuggestions/index.ts
@@ -1,0 +1,3 @@
+export * from './IApiSuggestionsProps';
+export * from './IApiSuggestionsState';
+export * from './ApiSuggestions';