+++ src/webparts/restTester/components/codeEditor/ICodeEditorProps.ts
@@ -1,0 +1,9 @@
+export interface ICodeEditorProps {
+  code: string;
+  language: string;
+  readOnly: boolean;
+  wordWrap: boolean;
+  height?: string;
+
+  onChange?: (val: string) => void;
+}