+++ src/webparts/restTester/components/headersInput/IHeadersInputProps.ts
@@ -1,0 +1,7 @@
+export interface IHeadersInputProps {
+  hIndex: number;
+  hKey: string;
+  hValue: string;
+
+  fUpdate: (i: number, key: string, value: string) => void;
+}