+++ src/webparts/restTester/components/headersInput/IHeadersInputState.ts
@@ -1,0 +1,4 @@
+export interface IHeadersInputState {
+  hKey: string;
+  hValue: string;
+}