+++ src/webparts/restTester/components/headersInput/index.ts
@@ -1,0 +1,3 @@
+export * from './IHeadersInputProps';
+export * from './IHeadersInputState';
+export * from './HeadersInput';