+++ src/webparts/restTester/components/responseInfo/ResponseInfo.tsx
@@ -1,27 +1,15 @@
 import * as React from 'react';
-import styles from './RestTester.module.scss';
+import styles from '../RestTester.module.scss';
 import { ActionButton } from 'office-ui-fabric-react/lib/components/Button';
 import { Checkbox } from 'office-ui-fabric-react/lib/Checkbox';
-import AceEditor from 'react-ace';
-import SnippetBuilder from './SnippetBuilder';
-import { ResultType, IRequestInfo } from './RestTester';
+import { SnippetBuilder } from '../snippetBuilder';
 import jsonToTS from 'json-to-ts';
 import { MessageBar, MessageBarType } from 'office-ui-fabric-react/lib/components/MessageBar';
+import { CodeEditor } from '../codeEditor';
+import { IResponseInfoProps, IResponseInfoState } from '.';
+import { ResultType } from '../RestTester';
 
-export interface IResponseInfoProps {
-  status: number | string;
-  resultType: ResultType;
-  wrapCode: boolean;
-  requestInfo: IRequestInfo;
-  data: string;
-
-  fSwitchTab: (val: ResultType) => void;
-  fTriggerCodeWrap: (ev: React.FormEvent<HTMLElement>, isChecked: boolean) => void;
-}
-
-export interface IResponseInfoState {}
-
-export default class ResponseInfo extends React.Component<IResponseInfoProps, IResponseInfoState> {
+export class ResponseInfo extends React.Component<IResponseInfoProps, IResponseInfoState> {
   public render(): React.ReactElement<IResponseInfoProps> {
     // Stringify the rest response
     const restResponse: string = this.props.data ? JSON.stringify(this.props.data, null, 2) : "";
@@ -61,34 +49,18 @@ export default class ResponseInfo extends React.Component<IResponseInfoProps, IR
 
         {
           this.props.resultType === ResultType.body && (
-            <AceEditor mode="json"
-                      theme="github"
-                      className={styles.codeZone}
-                      value={restResponse}
-                      readOnly={true}
-                      editorProps={{ $blockScrolling: true }}
-                      setOptions={{
-                        wrap: this.props.wrapCode,
-                        showPrintMargin: false,
-                        maxLines: restResponse ? restResponse.split(/\r\n|\r|\n/).length : 15
-                      }}
-                      width="100%" />
+            <CodeEditor language="json"
+                        code={restResponse}
+                        readOnly={true}
+                        wordWrap={this.props.wrapCode} />
           )
         }
         {
           this.props.resultType === ResultType.interface && (
-            <AceEditor mode="typescript"
-                      theme="github"
-                      className={styles.codeZone}
-                      value={interfaceObj}
-                      readOnly={true}
-                      editorProps={{ $blockScrolling: true }}
-                      setOptions={{
-                        wrap: this.props.wrapCode,
-                        showPrintMargin: false,
-                        maxLines: interfaceObj ? interfaceObj.split(/\r\n|\r|\n/).length : 15
-                      }}
-                      width="100%" />
+            <CodeEditor language="typescript"
+                        code={interfaceObj}
+                        readOnly={true}
+                        wordWrap={this.props.wrapCode} />
           )
         }
         {