+++ src/webparts/restTester/components/snippetBuilder/ISnippetBuilderProps.ts
@@ -1,0 +1,6 @@
+import { IRequestInfo } from "../RestTester";
+
+export interface ISnippetBuilderProps {
+  requestInfo: IRequestInfo;
+  wrapCode: boolean;
+}