+++ src/webparts/restTester/components/snippetBuilder/ISnippetBuilderState.ts
@@ -1,0 +1,3 @@
+export interface ISnippetBuilderState {
+  code: string;
+}