+++ src/webparts/restTester/components/snippetBuilder/index.ts
@@ -1,0 +1,3 @@
+export * from './ISnippetBuilderProps';
+export * from './ISnippetBuilderState';
+export * from './SnippetBuilder';